Reorganize auth package
