+++ app/src/main/res/layout/recycler_view_demo.xml
@@ -7,7 +7,7 @@
     tools:context=".RecyclerViewDemoActivity">
 
     <android.support.v7.widget.RecyclerView
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:id="@+id/messagesList"
         android:layout_alignParentLeft="true"
