+++ library/src/main/java/com/firebase/ui/auth/core/FirebaseLoginBaseActivity.java
@@ -1,4 +1,4 @@
-package com.firebase.ui;
+package com.firebase.ui.auth.core;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -7,12 +7,10 @@ import android.util.Log;
 import com.firebase.client.AuthData;
 import com.firebase.client.Firebase;
 import com.firebase.client.FirebaseError;
-import com.firebase.ui.auth.FacebookAuthHelper;
-import com.firebase.ui.auth.GoogleAuthHelper;
-import com.firebase.ui.auth.PasswordAuthHelper;
-import com.firebase.ui.auth.SocialProvider;
-import com.firebase.ui.auth.TokenAuthHandler;
-import com.firebase.ui.auth.TwitterAuthHelper;
+import com.firebase.ui.auth.facebook.FacebookAuthHelper;
+import com.firebase.ui.auth.google.GoogleAuthHelper;
+import com.firebase.ui.auth.password.PasswordAuthHelper;
+import com.firebase.ui.auth.twitter.TwitterAuthHelper;
 
 public abstract class FirebaseLoginBaseActivity extends AppCompatActivity {
