+++ library/src/main/java/com/firebase/ui/auth/core/FirebaseOAuthToken.java
@@ -1,4 +1,4 @@
-package com.firebase.ui.auth;
+package com.firebase.ui.auth.core;
 
 public class FirebaseOAuthToken {
     public String token;
