+++ library/src/main/java/com/firebase/ui/auth/core/FirebaseStatuses.java
@@ -1,4 +1,4 @@
-package com.firebase.ui.auth;
+package com.firebase.ui.auth.core;
 
 /**
  * Created by abehaskins on 11/9/15.
