+++ library/src/main/java/com/firebase/ui/auth/core/SocialProvider.java
@@ -1,4 +1,4 @@
-package com.firebase.ui.auth;
+package com.firebase.ui.auth.core;
 
 // Note: these enum values must be lowercase, to ensure they match the casing used by the Firebase Authentication providers
 public enum SocialProvider {
