+++ library/src/main/java/com/firebase/ui/auth/core/TokenAuthHandler.java
@@ -1,4 +1,4 @@
-package com.firebase.ui.auth;
+package com.firebase.ui.auth.core;
 
 import com.firebase.client.AuthData;
 import com.firebase.client.FirebaseError;
