+++ library/src/main/java/com/firebase/ui/auth/twitter/TwitterAuthHelper.java
@@ -1,12 +1,15 @@
-package com.firebase.ui.auth;
+package com.firebase.ui.auth.twitter;
 
 import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
-import android.util.Log;
 
 import com.firebase.client.Firebase;
 import com.firebase.client.FirebaseError;
+import com.firebase.ui.auth.core.FirebaseAuthHelper;
+import com.firebase.ui.auth.core.FirebaseOAuthToken;
+import com.firebase.ui.auth.core.FirebaseStatuses;
+import com.firebase.ui.auth.core.TokenAuthHandler;
 
 public class TwitterAuthHelper extends FirebaseAuthHelper {
